Added an income calculation on the death of an enemy.  This should add the income to the member's current amount.
To the instantiateEnemy method added a cost check to make sure the member has enough money to pay for the units.

Added a method that adds a new member to the enemyIncome list so they can start earning.

Added a cost and incomeMultiplier to the Enemy class.  Right now the incomeMultiplier is cost / 4f, this is used to determine how much money a member gets when an enemy dies.

Added bullet damage to enemies.

When a player kills an enemy he gains the cost of the enemy as a reward.